feat: Add Windows cross-compilation support and cleanup legacy code
- Added Windows cross-compilation support with conditional compilation
- Removed legacy Unix-only files (launcher_client.rs, monitor_server.rs)
- Fixed Windows-specific issues:
  - Path null terminator handling
  - .cmd file execution via cmd.exe wrapper
- Preserved full Unix functionality while adding Windows compatibility
- Updated dependencies with platform-specific configurations
- All tests pass, zero compiler warnings
