+++ launcher/src/tool_wrapper.rs
@@ -39,11 +39,23 @@ impl ToolWrapper {
 
     /// CLI ツール プロセスを起動（従来のパイプベース）
     pub async fn spawn(&self) -> Result<Child> {
-        let mut cmd = Command::new(self.tool.command_name());
-        cmd.args(&self.args);
+        let mut cmd = if cfg!(windows) {
+            // Windows環境では.cmdファイルを実行するためにcmd.exeを使用
+            let mut cmd = Command::new("cmd");
+            cmd.args(&["/C", self.tool.command_name()]);
+            cmd.args(&self.args);
+            cmd
+        } else {
+            let mut cmd = Command::new(self.tool.command_name());
+            cmd.args(&self.args);
+            cmd
+        };
 
         if let Some(working_dir) = &self.working_dir {
-            cmd.current_dir(working_dir);
+            // Windows環境でのnull terminator問題を回避
+            let path_str = working_dir.to_string_lossy();
+            let clean_path = path_str.trim_end_matches('\0');
+            cmd.current_dir(clean_path);
         }
 
         cmd.stdout(Stdio::piped())
@@ -67,14 +79,27 @@ impl ToolWrapper {
         let pty_pair = pty_system.openpty(get_pty_size())?;
 
         // コマンドを構築
-        let mut cmd = CommandBuilder::new(self.tool.command_name());
-        cmd.args(&self.args);
+        let mut cmd = if cfg!(windows) {
+            // Windows環境では.cmdファイルを実行するためにcmd.exeを使用
+            let mut cmd = CommandBuilder::new("cmd");
+            cmd.args(&["/C", self.tool.command_name()]);
+            cmd.args(&self.args);
+            cmd
+        } else {
+            let mut cmd = CommandBuilder::new(self.tool.command_name());
+            cmd.args(&self.args);
+            cmd
+        };
 
         // 作業ディレクトリを設定（指定がない場合は現在のディレクトリ）
         let working_dir = self.working_dir.clone().unwrap_or_else(|| {
             std::env::current_dir().unwrap_or_else(|_| std::path::PathBuf::from("."))
         });
-        cmd.cwd(working_dir);
+        
+        // Windows環境でのnull terminator問題を回避
+        let path_str = working_dir.to_string_lossy();
+        let clean_path = path_str.trim_end_matches('\0');
+        cmd.cwd(clean_path);
 
         // 共通環境変数を設定
         setup_common_pty_environment(&mut cmd);
