+++ monitor/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod live_ui;
-pub mod monitor_server;
 pub mod notification;
 pub mod session_manager;
 pub mod transport_server;
